requirements.txt
djangorestframework==3.12.4
djangorestframework_simplejwt==4.4.0
drf_yasg==1.20.0
python-dotenv==0.17.1
